refactor: カロリー入力フォームのコンポーネント化（手動部分のみ）（#5）
- CalorieInput: カロリー入力画面を表示する
- ManualCalorieForm: カロリー入力フォーム（手動）を表示する

※日付入力、エラーメッセージ、換算ボタンのコンポーネントは別コミットで管理済み
